refactor - moved methods to class. Add nickname
