+++ src/components/account/user.js
@@ -1,0 +1,6 @@
+import React from 'react';
+
+const SettingsUser = ({ user }) => {
+  return <div>Settings for customer</div>;
+};
+export default SettingsUser;
